+++ Security/db_dsmrapi.sql
@@ -19,4 +19,7 @@ GRANT EXECUTE ON [dbo].[DsmrApi_SelectPowerDataByHour] TO [db_dsmrapi]
 GO
 GRANT EXECUTE ON [dbo].[DsmrApi_SelectEnergyDataBetween] TO [db_dsmrapi]
 GO
-
+GRANT EXECUTE ON [dbo].[DsmrApi_ComputeDeviceCapabilities] TO [db_dsmrapi]
+GO
+GRANT EXECUTE ON [dbo].[DsmrApi_SelectAverageEnergyDataPerHour] TO [db_dsmrapi]
+GO