+++ CHANGELOG.md
@@ -4,7 +4,15 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [1.6.0] - 03-02-2021
+## [1.7.0] - 23-12-2021
+### Added
+- Stored procedure DsmrApi_ComputeDeviceCapabilities
+- Stored procedure DsmrApi_SelectAverageEnergyDataPerHour
+
+### Updated
+- DSMR API permissions
+
+## [1.6.0] - 06-06-2021
 ### Added
 - Permissions for the DSMR API
 - Stored procedure to compute energy usage/production